+++ NEWS.md
@@ -1,3 +1,10 @@
+# microViz (development version)
+
+## Features
+Robust clr and robust aitchison distance are now supported
+- `tax_transform` supports "rclr" (robust centered log ratio)
+- `dist_calc` supports "robust.aitchison" distance (as rclr & euclidean distance)
+
 # microViz 0.9.2
 
 ## Features